changing group size to size_t

+++ src/plugins/intel_gpu/include/intel_gpu/op/dynamic_quantize.hpp
@@ -19,15 +19,16 @@ public:
     /// \brief Constructs an DynamicQuantize operation.
     ///
     /// \param data Input tensor with data
-    DynamicQuantize(const Output<Node>& data, int64_t group_size);
+    /// \param group_size Group size for dynamic quantization
+    DynamicQuantize(const Output<Node>& data, size_t group_size);
 
     void validate_and_infer_types() override;
 
     std::shared_ptr<Node> clone_with_new_inputs(const ov::OutputVector& new_args) const override;
-    int64_t get_group_size() { return m_group_size; };
+    size_t get_group_size() { return m_group_size; };
 
 private:
-    int64_t m_group_size;
+    size_t m_group_size;
 };
 
 std::vector<ov::PartialShape> shape_infer(const DynamicQuantize* op, std::vector<ov::PartialShape> input_shapes);
